requirements.txt
alembic==1.0.1
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3
cssselect==1.0.3
enum34==1.1.6
ERAlchemy==1.2.10
fake-useragent==0.1.11
graphviz==0.9
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipaddress==1.0.22
lxml==4.2.4
Mako==1.0.7
MarkupSafe==1.0
parsel==1.5.0
Pillow==5.3.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.18
PyDispatcher==2.0.5
pygraphviz==1.3.1
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
queuelib==1.5.0
requests==2.21.0
Scrapy==1.5.1
scrapy-crawlera==1.5.0
scrapy-fake-useragent==1.1.0
scrapyd==1.2.0
scrapyd-client==1.2.0a1
selenium==3.14.0
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.10
Twisted==18.7.0
urllib3==1.23
w3lib==1.19.0
zope.interface==4.5.0
